✨ pass EvmCanisterService to functions
