+++ packages/evm-rpc-canister-types/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "evm-rpc-canister-types"
-version = "0.1.1"
+version = "0.1.2"
 edition = "2021"
 readme = "README.md"
 authors = ["Moritz Fuller moritz.fuller@dfinity.org"]
