+++ packages/ic-evm-utils/src/request.rs
@@ -1,4 +1,4 @@
-use evm_rpc_canister_types::{RequestResult, RpcService, EVM_RPC};
+use evm_rpc_canister_types::{EvmRpcCanister, RequestResult, RpcService};
 use serde::{Deserialize, Serialize};
 
 use crate::request_cost::request_cost;
@@ -7,16 +7,18 @@ pub async fn request(
     rpc_service: RpcService,
     json_rpc_payload: String,
     max_response_bytes: u64,
+    evm_rpc: EvmRpcCanister,
 ) -> RequestResult {
     // estimate cycles costs
     let cycles = request_cost(
         rpc_service.clone(),
         json_rpc_payload.clone(),
         max_response_bytes,
+        evm_rpc.clone(),
     )
     .await;
     // call request with estimated cycles
-    match EVM_RPC
+    match evm_rpc
         .request(rpc_service, json_rpc_payload, max_response_bytes, cycles)
         .await
     {
